migrate singular and plural data source to new SDK
