+++ internal/service/federatedsettingsidentityprovider/data_source_federated_settings_identity_providers.go
@@ -4,14 +4,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"sort"
-	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/conversion"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
-	matlas "go.mongodb.org/atlas/mongodbatlas"
 )
 
 func PluralDataSource() *schema.Resource {
@@ -216,214 +212,26 @@ func PluralDataSource() *schema.Resource {
 }
 func dataSourceMongoDBAtlasFederatedSettingsIdentityProvidersRead(ctx context.Context, d *schema.ResourceData, meta any) diag.Diagnostics {
 	// Get client connection.
-	conn := meta.(*config.MongoDBClient).Atlas
+	connV2 := meta.(*config.MongoDBClient).AtlasV2
 
 	federationSettingsID, federationSettingsIDOk := d.GetOk("federation_settings_id")
-
-	options := &matlas.ListOptions{
-		PageNum:      d.Get("page_num").(int),
-		ItemsPerPage: d.Get("items_per_page").(int),
-	}
 
 	if !federationSettingsIDOk {
 		return diag.FromErr(errors.New("federation_settings_id must be configured"))
 	}
+	// There is no available paginated method
+	// If no protocol is specified, only SAML identity providers will be returned. Use ListIdentityProvidersWithParams to specify protocol
+	federatedSettingsIdentityProviders, _, err := connV2.FederatedAuthenticationApi.ListIdentityProviders(ctx, federationSettingsID.(string)).Execute()
 
-	federatedSettingsIdentityProviders, _, err := conn.FederatedSettings.ListIdentityProviders(ctx, federationSettingsID.(string), options)
 	if err != nil {
 		return diag.Errorf("error getting federatedSettings IdentityProviders assigned (%s): %s", federationSettingsID, err)
 	}
 
-	if err := d.Set("results", flattenFederatedSettingsIdentityProvider(federatedSettingsIdentityProviders)); err != nil {
+	if err := d.Set("results", FlattenFederatedSettingsIdentityProvider(federatedSettingsIdentityProviders)); err != nil {
 		return diag.FromErr(fmt.Errorf("error setting `result` for federatedSettings IdentityProviders: %s", err))
 	}
 
 	d.SetId(federationSettingsID.(string))
 
 	return nil
-}
-
-func flattenFederatedSettingsIdentityProvider(federatedSettingsIdentityProvider []matlas.FederatedSettingsIdentityProvider) []map[string]any {
-	var federatedSettingsIdentityProviderMap []map[string]any
-
-	if len(federatedSettingsIdentityProvider) > 0 {
-		federatedSettingsIdentityProviderMap = make([]map[string]any, len(federatedSettingsIdentityProvider))
-
-		for i := range federatedSettingsIdentityProvider {
-			federatedSettingsIdentityProviderMap[i] = map[string]any{
-				"acs_url":                      federatedSettingsIdentityProvider[i].AcsURL,
-				"associated_domains":           federatedSettingsIdentityProvider[i].AssociatedDomains,
-				"associated_orgs":              flattenAssociatedOrgs(federatedSettingsIdentityProvider[i].AssociatedOrgs),
-				"audience_uri":                 federatedSettingsIdentityProvider[i].AudienceURI,
-				"display_name":                 federatedSettingsIdentityProvider[i].DisplayName,
-				"issuer_uri":                   federatedSettingsIdentityProvider[i].IssuerURI,
-				"okta_idp_id":                  federatedSettingsIdentityProvider[i].OktaIdpID,
-				"pem_file_info":                flattenPemFileInfo(*federatedSettingsIdentityProvider[i].PemFileInfo),
-				"request_binding":              federatedSettingsIdentityProvider[i].RequestBinding,
-				"response_signature_algorithm": federatedSettingsIdentityProvider[i].ResponseSignatureAlgorithm,
-				"sso_debug_enabled":            federatedSettingsIdentityProvider[i].SsoDebugEnabled,
-				"sso_url":                      federatedSettingsIdentityProvider[i].SsoURL,
-				"status":                       federatedSettingsIdentityProvider[i].Status,
-			}
-		}
-	}
-
-	return federatedSettingsIdentityProviderMap
-}
-
-func flattenAssociatedOrgs(associatedOrgs []*matlas.AssociatedOrgs) []map[string]any {
-	var associatedOrgsMap []map[string]any
-
-	if len(associatedOrgs) == 0 {
-		return nil
-	}
-	associatedOrgsMap = make([]map[string]any, len(associatedOrgs))
-
-	for i := range associatedOrgs {
-		if associatedOrgs[i].UserConflicts == nil {
-			associatedOrgsMap[i] = map[string]any{
-				"domain_allow_list":          associatedOrgs[i].DomainAllowList,
-				"domain_restriction_enabled": associatedOrgs[i].DomainRestrictionEnabled,
-				"identity_provider_id":       associatedOrgs[i].IdentityProviderID,
-				"org_id":                     associatedOrgs[i].OrgID,
-				"post_auth_role_grants":      associatedOrgs[i].PostAuthRoleGrants,
-				"role_mappings":              FlattenRoleMappings(associatedOrgs[i].RoleMappings),
-				"user_conflicts":             nil,
-			}
-		} else {
-			associatedOrgsMap[i] = map[string]any{
-				"domain_allow_list":          associatedOrgs[i].DomainAllowList,
-				"domain_restriction_enabled": associatedOrgs[i].DomainRestrictionEnabled,
-				"identity_provider_id":       associatedOrgs[i].IdentityProviderID,
-				"org_id":                     associatedOrgs[i].OrgID,
-				"post_auth_role_grants":      associatedOrgs[i].PostAuthRoleGrants,
-				"role_mappings":              FlattenRoleMappings(associatedOrgs[i].RoleMappings),
-				"user_conflicts":             FlattenUserConflicts(*associatedOrgs[i].UserConflicts),
-			}
-		}
-	}
-
-	return associatedOrgsMap
-}
-
-func FlattenUserConflicts(userConflicts matlas.UserConflicts) []map[string]any {
-	var userConflictsMap []map[string]any
-
-	if len(userConflicts) == 0 {
-		return nil
-	}
-	userConflictsMap = make([]map[string]any, len(userConflicts))
-
-	for i := range userConflicts {
-		userConflictsMap[i] = map[string]any{
-			"email_address":          userConflicts[i].EmailAddress,
-			"federation_settings_id": userConflicts[i].FederationSettingsID,
-			"first_name":             userConflicts[i].FirstName,
-			"last_name":              userConflicts[i].LastName,
-			"user_id":                userConflicts[i].UserID,
-		}
-	}
-
-	return userConflictsMap
-}
-
-func flattenPemFileInfo(pemFileInfo matlas.PemFileInfo) []map[string]any {
-	var pemFileInfoMap []map[string]any
-
-	if len(pemFileInfo.Certificates) > 0 {
-		pemFileInfoMap = make([]map[string]any, 1)
-
-		pemFileInfoMap[0] = map[string]any{
-			"certificates": flattenFederatedSettingsCertificates(pemFileInfo.Certificates),
-			"file_name":    pemFileInfo.FileName,
-		}
-	}
-
-	return pemFileInfoMap
-}
-
-func flattenFederatedSettingsCertificates(certificates []*matlas.Certificates) []map[string]any {
-	var certificatesMap []map[string]any
-
-	if len(certificates) > 0 {
-		certificatesMap = make([]map[string]any, len(certificates))
-
-		for i := range certificates {
-			certificatesMap[i] = map[string]any{
-				"not_after":  conversion.TimeToString(certificates[i].NotAfter),
-				"not_before": conversion.TimeToString(certificates[i].NotBefore),
-			}
-		}
-	}
-
-	return certificatesMap
-}
-
-type mRoleAssignment []*matlas.RoleAssignments
-
-func (ra mRoleAssignment) Len() int      { return len(ra) }
-func (ra mRoleAssignment) Swap(i, j int) { ra[i], ra[j] = ra[j], ra[i] }
-func (ra mRoleAssignment) Less(i, j int) bool {
-	compareVal := strings.Compare(ra[i].OrgID, ra[j].OrgID)
-
-	if compareVal != 0 {
-		return compareVal < 0
-	}
-
-	compareVal = strings.Compare(ra[i].GroupID, ra[j].GroupID)
-
-	if compareVal != 0 {
-		return compareVal < 0
-	}
-
-	return ra[i].Role < ra[j].Role
-}
-
-type roleMappingsByGroupName []*matlas.RoleMappings
-
-func (ra roleMappingsByGroupName) Len() int      { return len(ra) }
-func (ra roleMappingsByGroupName) Swap(i, j int) { ra[i], ra[j] = ra[j], ra[i] }
-
-func (ra roleMappingsByGroupName) Less(i, j int) bool {
-	return ra[i].ExternalGroupName < ra[j].ExternalGroupName
-}
-
-func FlattenRoleMappings(roleMappings []*matlas.RoleMappings) []map[string]any {
-	sort.Sort(roleMappingsByGroupName(roleMappings))
-
-	var roleMappingsMap []map[string]any
-
-	if len(roleMappings) > 0 {
-		roleMappingsMap = make([]map[string]any, len(roleMappings))
-
-		for i := range roleMappings {
-			roleMappingsMap[i] = map[string]any{
-				"external_group_name": roleMappings[i].ExternalGroupName,
-				"id":                  roleMappings[i].ID,
-				"role_assignments":    FlattenRoleAssignments(roleMappings[i].RoleAssignments),
-			}
-		}
-	}
-
-	return roleMappingsMap
-}
-
-func FlattenRoleAssignments(roleAssignments []*matlas.RoleAssignments) []map[string]any {
-	sort.Sort(mRoleAssignment(roleAssignments))
-
-	var roleAssignmentsMap []map[string]any
-
-	if len(roleAssignments) > 0 {
-		roleAssignmentsMap = make([]map[string]any, len(roleAssignments))
-
-		for i := range roleAssignments {
-			roleAssignmentsMap[i] = map[string]any{
-				"group_id": roleAssignments[i].GroupID,
-				"org_id":   roleAssignments[i].OrgID,
-				"role":     roleAssignments[i].Role,
-			}
-		}
-	}
-
-	return roleAssignmentsMap
 }
